Move some of the complicated session averaging logic to GameSummary model, make it slightly less dumb of a struct, so GameSummary resource has less responsibility.
